task1/requirements.txt
SpeechRecognition==3.9.0
pyttsx3==2.90
pyaudio==0.2.13
requests==2.31.0
feedparser==6.0.10
python-dateutil==2.8.2
schedule==1.1.0
